More Barsigns, Service Holosign (#20963)
After a brief discourse in Discord #general channel it was decided to
add some more barsigns and a new one (I had sitting around for 500 years
on my PC). Three new barsigns were added:

![image](https://github.com/user-attachments/assets/d8a4f87d-5cf2-458a-b584-a397a985db2a)

With the following descriptions:

![image](https://github.com/user-attachments/assets/4c1d5893-2504-4174-8066-d593259a1e2c)


The new holosign looks like this:

![image](https://github.com/user-attachments/assets/2f3ecd5d-5dc7-4768-8c08-500a4388d6ea)

Has a nice little animation. Signs and buttons are mapped into bar and
kitchen. Some really minor shuffling was done for the signs. Tested
locally and it works on my machine (💯).

Please enjoy.
### Asset Licenses
The following assets that **have not** been created by myself are
included in this PR:

| Path | Original Author | License |
| --- | --- | --- |
| barsigns.dmi| Unknown (BayStation12) | AGPL v3.0 |

---------

Signed-off-by: KingOfThePing <43940569+KingOfThePing@users.noreply.github.com>
Co-authored-by: SleepyGemmy <99297919+SleepyGemmy@users.noreply.github.com>