Changes the name used in the code for the solarian army beret (#21028)
Changes the last part of the name used in the code for the solarian army
beret from "beret/sapk" to "beret/sol/army".

Most clothing objects associated with the Solarian military have "sol"
in their name somewhere and I think it would be good for this one to as
well so it shows up alongside all the other ones when you look up "sol"
in, for example, the item spawner.

---------

Signed-off-by: Tag103 <194812614+Tag103@users.noreply.github.com>