Replaces the Solarian Marine event outfit's energy sword with a solarian energy dagger (#21027)
Replaces the energy sword in the belt of the Solarian Marine event
outfit with a solarian energy dagger.

The outfit was created before the solarian energy dagger was added and I
believe it otherwise would have used it as its melee weapon in place of
the energy sword, since it's supposed to be standard issue for the
Solarian military.

The Heavy Solarian Marine outfit is not changed by this and still has
the sword beacuse I think it is fitting for the heavy variant of the
outfit to have a bigger, stronger and less common melee weapon than the
regular one.

---------

Signed-off-by: Tag103 <194812614+Tag103@users.noreply.github.com>