The Great Sprite Unification: Part 1 (#21006)
In an effort to make the way our sprites are organized make a little bit
more sense, this takes ALL the sprites in `icons/clothing` and moves
them over `icons/obj/item/clothing` or another appropriate file in the
case of non-clothing items.

A lot of these files had to be split since they lumped everything into
the same dmi which had to be changed since people, reasonably, expect to
find a hat in the folder named `hat` and not in the one where we're
supposed to keep uniforms in.

A lot of smaller files were also merged into already existing dmis where
it made sense.

Lastly, all the colourable items kept in `icons/obj/item/clothing` were
updated to the new palette. Those items were:

**Backpacks:**
- GD Sacred Icon

**Belts:**
- Fannypack

**Gloves:**
- Gloves
- Single Glove

**Hats:**
- GD Cowl
- GD Cube Hood
- Furred Crown
- Surgeon Cap
- Kippah
- Plain Hood

**Shoes:**
- Konyanger Gomusin

**Coats:**
- GD Eccentric Coat

**Uniforms:**
- Colourable Dress
- Shortsleeve Dress
- Evening Gown
- Open Shoulder Dress
- Assymetric Dress
- Tea Dress
- Long Sleeve Dress
- Colourable Scrubs
- Colourable Jumpsuit