The Bodega Update: remapping the commissary and library (#21029)
This PR remaps the operations commissary, moving it across to the
service hallway where it can receive much more foot traffic than in its
original position. Reflecting its new spot in service, and considering
particularly how much a hydroponicist can contribute to the commissary,
both service and operations staff now have access to it - although I
personally anticipate this will remain primarily used by operations.

Adjacent vendors in the service hall have been moved into the commissary
to encourage interaction, and hopefully to capture a little of that
bodega feel - although I'm open for other ideas on how to capture the
right atmosphere.

Because of the location, the library has necessarily also been remapped.
It's now smaller, but my hope is that it uses the space it has more
effectively. The old location of the commissary is now generic
maintenance until someone finds a better use for it.

A few points:

- "Is this bad for librarians?" Hopefully not! It isn't a very popular
job right now, and I don't expect this PR to change that, but I've gone
into the remap with the intention of making the library a more appealing
place to be in than before. The old library has a lot of dead space,
particularly with its gigantic office, whereas this new design is
intended to be a little more compact. Now featuring tinted windows if
you want to get some mood lighting going on.
- "Why the move?" The current location of the commissary is within view
of virtually no foot traffic period - operations generally is an
extremely isolated department, and something like the commissary lives
and dies in the playerbase according to how visible it is.
Appropriately, this moves it to a more visible spot.

<img width="416" height="544" alt="image"
src="https://github.com/user-attachments/assets/c470abfa-a01c-467c-8fff-58a308b0531c"
/>

<img width="480" height="256" alt="image"
src="https://github.com/user-attachments/assets/aa32a756-2573-4746-bebf-9002e4f66cd9"
/>