Refactor Brain Math (#20711)
This PR fixes an issue where characters were experiencing brain damage
at rates that were sometimes inconsistent depending on player population
and server performance. To accomplish this, it makes it so that brain
damage taken is no longer a random amount per second that fluctuates
based on server performance, and is instead an amount of damage per
second that scales with the previous conditions, but is "tick
invariant", meaning it no longer changes if for instance you have 75
players online causing server lag.

The benefit of this is particularly noticeable in the extreme ends of
lowpop as well as high pop. It'll no longer cause brain-damage at
hyperspeed if theres' only 10 people online.