Fixes vending machine overlays when losing power (#21062)
Fixes https://github.com/Aurorastation/Aurora.3/issues/21023

So apparently the overlays were getting reapplied when the vending
machines lost power. This solves the issue for both the coffee vending
machine and other vending machines that had overlays, such as the idris
soda machine and the booze-o-mat, to name a few.

---------

Signed-off-by: Casper3667 <8396443+Casper3667@users.noreply.github.com>
Co-authored-by: Wowzewow (Wezzy) <42310821+alsoandanswer@users.noreply.github.com>