Fix phazon mech trampling itself (#21071)
Observed in the jockey round earlier (cAO-dL7N) that a phazonned mech
was trampling and destroying itself.

I couldn't reproduce on the dev map with a modified
`/mob/living/heavy_vehicle/premade/random`, but could in the jockey
antag area with daemon mech parts. In game they were still trampling
themselves once out of phazon mode, but I was unable to replicate that.
This fix should handle that case, still.

Also observed the phazonned mech being attackable, though I'm not sure
that's strictly a bug so I'm leaving it alone in favor of fixing the
biggest issue at the moment.