Adds the security chestrig to the warden, HoS and sec officer lockers, as an alternative to the security belt (#21058)
As the title says. The PR was done on the request of VeteranGary with
sprites provided by Noble Row.
The reasoning for it is to give some more visual variety, by allowing it
to appear over the standard security armor plate. It has been added as a
belt for that purpose, with the reasoning that there is the explorer's
belt that similarly is a chestrig.

### Asset Licenses
The following assets that **have not** been created by myself are
included in this PR:

| Path | Original Author | License |
| --- | --- | --- |
| icons/mob/belt.dmi | Noble Row (Aurora Station) | CC0 |
| icons/obj/clothing/belts.dmi | Noble Row (Aurora Station) | CC0 |