Michael Wave, Return of Microwaves (#21072)
- Readds microwaves, and most of their old recipes where applicable.
- Resprites the small oven. It was small and ugly and I hated it.

Now, you may be asking, Wezzy - some of these recipes are an affront to
god. Microwaves can't do that.
And to that I say, It's mostly there to be a decent stand-in for small
rinky-dink maps and ships where you can't particularly fit the whole
kitchen suite in a convenient place. It's nice to have a bit of an
asspull in places like these, so.

Also, donk pockets actually become hot and cool down properly again,
now. It only took 15 years. (jesus christ)

Ported from Bay which ported from us.


![dreamseeker_6aMUIkYymn](https://github.com/user-attachments/assets/76c64677-78b2-4a3b-a35a-002b1cfe7cc7)