Exo-stellar skeletons now start with a storage module (#21089)
Title. Giving up a backpack slot for what is essentially a fluff item is
a pretty big ask. The only real advantage it has over just popping an
RMT pill is the fact it has welding protection, which the people who
typically need it can already get from other items, and that it can
inject dexalin which, although it might come in handy in a pinch,
doesn't justify massively gimping your storage capacity for.
The storage module is still significantly worse than a regular backpack
so there are still some drawbacks to bringing the exoskeleton, but not
the extreme level that it sits at now.