All hardsuits remove Offworlder slowdown (#21091)
Currently, picking the exostellar skeleton as an offworlder means you
can't use a hardsuit since only the ESS removes the slowdown. This fixes
that by extending that property to all other hardsuits with the logic
that if they're strong enough to splint a miner's leg and let let them
walk around, they're strong enough to help noodleman with their back
pain.