Mech Quality Of Life Additions (#21081)
This PR adds a few much needed new features to mechs, with the main one
being Mechanics Hints. Mechs have a unique way of handling mechanical
hints, in that shift-clicking on a mech to get these hints will reveal
the hints for its equipped modules.

I've also gone through and eliminated all of the lag causing hard
deletes present in the mech code. In a typical round these won't occur
all that often, but if I'm touching all these files I might as well do
my part and certify them free of hard deletes.

Lastly, after talking with Shaft Mining players, I discovered that the
main reason they don't ever want to use mining mechs is that the warp
extraction bag is strictly superior to mining with a mech, as it lets
them teleport ores directly to the ship. So I've added an option to link
ore summoners directly to these "Warp crates", allowing a mech equipped
with one to teleport ores to the ship. Just like the bag can.