Remove weakened hivebot beacon from golden slime (#20163)
In round cwE-dDLE a "dilapidated hivebot beacon" was summoned by gold
slime cores. It was invisible and continuously summoned hivebots. Since
the other types of hivebot beacons are already blacklisted, this
prevents this type of hivebot beacon from being summoned by the gold
slime cores as well. I assume it got overlooked because it's in another
file - there's no reason for a single use item to spawn an infinitely
free hivebot summoner. Another PR can fix the invisibility issue, if its
cause is discovered and it's not a gold core specific interaction
causing it.

For ease of reference, the weakened hivebot beacon is here:

https://github.com/Aurorastation/Aurora.3/blob/6d7c058ee7b44ba73099f46116cbd819fb86cba2/maps/away/away_site/hivebot_hub/hivebot_hub.dm#L119