Add pre-filled belts to the investigator lockers (#21128)
As security officers have their belts pre-filled, I thought to do the
same for the investigators. This takes the tape, flash and camera from
their locker and adds it to the belt in the same locker, additionally a
tape recorder is also put in it.
As clipboards can be put in the belts, I did find it weird that the
notepads couldn't, so I also added the ability to put notepads into the
security belt, as it felt fitting to be able to do so.
The extra pair of identical shoes were also removed from the locker, to
help reduce the clutter slightly. One pair still spawns in it.