Adds Offworlder HOS (#21088)
as the title says.

in my long contemplation of adding this, people have offered arguments
that an offworlder hos would be unable to fit the requirements for a
security job. i disagree here because not only do they have a lower
brute modifier than M'sai Tajara (who are the only tajara species
allowed to be HOS), but the negatives which would prevent them from
fitting any requirements can easily be made up with the various items
available to offworlders which pretty much alleviate any concerns here.
additionally, given off-worlders are currently allowed to be other
physically strenuous jobs which require even more exertion than HOS
would, such as paramedic, atmos tech, miner, or even security officers
(while i understand some think the final job is also undesirable, as far
as i am aware it has created no issues in the past), there seems to be a
massive inconsistency in the standards applied on offworlders.

so, i think that allowing offworlders to be HOS on the ship does make
sense, as they are adapted for work in space, and allows for a few cool
potential character concepts, with any drawbacks being pretty
inconsequential.