Space heater crash safety (#21182)
Space heaters, when spawned via admin panel, will crash if if their TGUI
is accessed immediately. This is because var/env is only initialized on
process(), and if not initialized, the game will be very sad when the UI
tries to display the current temperature by rounding a null value.

Sets env on Initialize() and cleans up /ui_act() logic to prevent crash.

---------

Signed-off-by: Batrachophreno <Batrochophreno@gmail.com>