gear loadout price standardization (#21159)
* "Lowers higher than usual prices of NT, Imperial army and TCAF
gearloadouts."

Some of the gear loadouts had higher than usual prices of 50 per group
and 15 per single compared to the standard 35 and 10. Some of these like
TCAF even were just straight up the same items as the single crate but
at less bang per buck.

Overall none of these kits are strong enough when compared to others, to
justify a higher price that makes it hard for mercs to budget it for
their gimmick. While the NT kit has a hardsuit, it should be noted it is
an empty one that has online slowdown unlike other combat rigs, similar
to the TCFL rigs.

Co-authored-by: Big Rat <lienesica@gmail.com>