Gun Sound Reminder (#21193)
## About PR

In preparation for more new gun sounds, this PR simply adds a header to
the boxes.dm, which handles reload sounds for all types of guns to
remind contributors about the existence of the different types of sounds
and when they could/should be used.
The way the game handles reloads the sounds are tied to the magazines.
While testing something completely unrelated I noticed that a lot guns
recently added use the standard reload sound, instead of a maybe more
appropriate one.

No player facing changes.