Fixes mobs ignoring their speed value (#21203)
## About PR

the `delay` value in use here gets assigned with mob's `speed` var (I am
positive on this, though movement code is a relic I can't fully
comprehend). If the mob calling this proc had an initial delay value,
this check used to ignore it and assigned a fix number. This PR makes it
so if `speed` is null or 0, then they're assigned with the fix number.