Give officers and wardens breath analyzers and investigators a gas mask (#21215)
From a request I saw a bit ago, I have now added breath analyzers to the
security officer and warden lockers, should they wish to use them. It is
mostly intended to check if people are drunk on duty or similar.

Investigators were given a half gas mask, given the amount of
pepperspray used in some cases and they were the only ones not to get it
in the department. If they want the bigger mask, they can ask the
warden.

Lastly while I was messing around in the file, I cleaned out the cadet
and investigator locker slightly. The clothing items that already
existed in the security drobe were removed from their lockers. The
security officer and warden lockers were already free of that, so they
did not get anything removed.