Nightshade injectors now cost TC (#21183)
* Berserk injector cost changes from 4 bluecrystals to 3 telecrystals.
Bluecrystals tend to be for defensive, support, utility or roleplay
items. Red Nightshade is for the most part a combat item. So this
changes its cost to be in TC, slightly lowering it to match the higher
value currency being used, since it is in the end also still a limited
use item that damages your organs.

---------

Signed-off-by: UltraBigRat <111320234+UltraBigRat@users.noreply.github.com>
Co-authored-by: Big Rat <lienesica@gmail.com>