The SIIB - Outfits and Weapons (#21007)
The Solarian Interstellar Intelligence Bureau is added, including an
Agent (Undercover) and Operative (Overt) variant.

- Adds in 2 new sprites for the SIIB based off the Solarian SMG and
Pistol
- Adds in 2 new outfits for Admins to play around with during events

These are just another thing for admins to mess around with, either for
Admins or otherwise.

Any other guidance on modifications to make this more lore-accurate
would be nice, alongside anything else to add.