More Miscellaneous Signs, Part I (#21205)
## About PR
This PR adds a few miscellaneous directional and otherwise useful signs,
both for the Horizon and potential away sites. The new ones can be seen
in the screenshot below. It also enables the use of one already existing
but not coded in sign.
As the title implies more is on the way. But for now, please enjoy.

> [!IMPORTANT]
> None of the signs have been added yet to the Horizon or elsewhere. 😀

>  Addition to signs.dmi:
> <img width="1081" height="212" alt="image"
src="https://github.com/user-attachments/assets/d5d9bca3-48c1-45b5-8453-59ee0495c96e"
/>