Gives hardsuits to group loadouts that were missing them. (#21207)
* Alot of newer group loadouts were missing their factions hardsuit,
despite that being the standard for most of them, so this gives those to
the faction loadouts that had appropriate ones.
* Also gave the Solarian hardsuit loadout a version with modules since
it was an empty version unlike the other hardsuits previously. Used the
fsf module set for that since their regular equipped version had a pulse
rifle module for some reason, suit itself should still be the same.

---------

Co-authored-by: Big Rat <lienesica@gmail.com>