Adds thermals and binoculars to the cultist spawn areas on the cult base (#21244)
The purpose of this PR isn't to make the cultists stronger in combat but
rather to give them the tools to better track whoever's exploring the
site. This gives the cultists a better opportunity to 'stalk' the
explorers and actually do spooky things out of their sight, like leaving
stuff around for them to find and just generally build up the eventual
encounter better, as opposed to just stumbling into them.

To help facilitate that idea I also mapped in some paper bins & pencils
throughout the map so the cultists can write their own spooky notes if
they want.