Himean voidsuits sprite uplift (#21247)
With this PR I will finally show you all that I am literally
**Him**(ean).
Okay but this PR is meant to be a bit of a dedication/advertisement to
what the Taj lore team has been cooking with the Free Tajaran Council
and the FREE/TC/.HIMEO. Do read it as it's quite neat, and make a Orion
Express FTC tajara miner today! (this is a threat)

Okay, but PR details now, basically, tiny uplift of Himean voidsuits and
their Tajara variants, that's about it, but, the changelog says it best:
> imageadd: "Resprited the himean voidsuits and its tajaran variants."
> rscdel: "Removed the mention of a flagpatch on the himean voidsuits to
reflect the appearance of the actual sprites."

Images, as always, in the details below
Old (current) sprites on the top, new on the bottom.

<details>
<summary>All the voidsuits resprited in this PR: </summary>
<img width="3073" height="1537" alt="literally him(ean)"
src="https://github.com/user-attachments/assets/1a3f548c-25da-44f7-8563-4aa6a74c5786"
/>

</details>

And, of course,
Punch your local Hephaestus Industries representative today!