Revamped lottery cards (#21261)
**Daddy needs a new pair of shoes!**

Revamped the lottery cards to both not reward astronomical amounts of
money anymore, and resprites them to make them look prettier and more
"digital" in line with how the card itself is basically an imitation
scratcher you shove into an ATM later on anyway.

<img width="301" height="184" alt="image"
src="https://github.com/user-attachments/assets/48e4c8c9-b783-413d-89d8-779a2538134c"
/>

* Gray: Bust
* Green: Slight win (1x to 1.25x)
* Blue: Big win (2-5x)
* Purple: Huge win (50-100x)
* Gold: Jackpot (1000x)

After three scratches, the card becomes available to redeem in an ATM or
in anything that accepts charge cards or IDs (it is basically a charge
card with an actual balance). I wouldn't exactly sanction buying a soda
with a lottery ticket though.

## The Mechanics
The lotto card (now selling for 10电 in the cigarette machine) plays
best-of-3 with 3 scratches per card. Essentially this means only the
highest prize will be kept; if you for example win $10 twice on the same
card, you will only win $10. This is the **house edge.** Rules of the
game, bucko.

Available prizes and odds are (for the 10电 variant) **PER SCRATCH** are:
* Loss: 89.495%
* Break-even (1x): 8.7% (10电)
* Small win (1.25x): 1.15% (12电)
* Double win (2x): 0.48% (20电)
* 5x win: 0.095% (50电)
* 50x win: 0.05% (500电)
* 100x win: 0.02% (1,000电)
* JACKPOT (1000x): 0.0099% (10,000电)
* SUPER JACKPOT (10,000x): 0.0001% (100,000电)

For per ticket odds (best of 3):

- Lose all: ~71.7%
- Top prize = 1× (break-even): ~23.0%
- 1.25x: ~3.4%
- 2x: ~1.4%
- 5x: ~0.28%
- 50x: ~0.15%
- 100x: ~0.06%
- Jackpot (1,000x): ~0.027% (≈ 1 in 3,367 tickets)
- Super Jackpot (10,000x): ~0.0003% (≈1 in 333,333 tickets)

Will you lose spectacularly? Probably. But 99.9% of gamblers quit before
they hit the top. That 100,000 credit big prize _calls_ to you...

Just one more pull.