Adds the ability for factions to blacklist specific citizenships (#21129)
Nothing player-facing in this one, for now. This adds the ability for
factions to be unavailable for certain citizenships.
It has been implemented as a blacklist due to the variety of
citizenships in the game. An example would be if the DPRA was added to
the list for Idris, then anybody with a DPRA citizenship would be unable
to select Idris as their faction.
This was done for the taj lore team.

---------

Signed-off-by: Casper3667 <8396443+Casper3667@users.noreply.github.com>
Co-authored-by: SleepyGemmy <99297919+SleepyGemmy@users.noreply.github.com>