Maps In The New Command Noticeboard (#21336)
## About PR

With the new noticeboards added the new subtype is now being mapped in:
> <img width="1362" height="1136" alt="image"
src="https://github.com/user-attachments/assets/22455a77-f849-45ca-94cd-29f5e869b256"
/>

It is placed centrally (so everyone can look at it). To unlock the
maglock of the protective glass (not shown in picture) a command level
access ID is needed. As soon as it is unlocked it behaves like regular
cabinets. It is meant for command to pin up important public notices. A
reminder that the command noticeboard has a limit of six (6) notices.
Please enjoy.