Minor tweaks to Freebooter ship cost (#21306)
https://github.com/Aurorastation/Aurora.3/blob/e1226fcd7240bef8b80e0683fd7e58437d0e6927/maps/_common/mapsystem/map.dm#L309-L318

Ship cost doesn't play a role here when a map is picked among the
candidates pool. There is no reason to not treat a variation ship as a
full ship after it's chosen. This was leaving budget points as a decimal
number, preventing the loop from ending early before eliminating each
candidates.

This doesn't change the outcome but makes it more in line with the
intended standard. I had no idea about this at the time when I made
freebooter variation (oops)