Adds the price and crate scanner to the hangar tech lockers and toolbelt (#21300)
This adds the price scanner, crate scanner and a clipboard to the hangar
technician lockers.
It also add the various scanners alongside the quikpay to the toolbelt,
as it could already have the EFTPOS scanner.
Lastly while doing this, I removed some of the department clothes from
the lockers that were already in their drobe.