Readds Kathira El-Hashem's custom items (#21354)
After discussing it with WickedCybs, it was permitted for this to be
readded now that I once more have the WL and can play Kathira again.
The edge of the cloak, alongside the journal had sprite updates to bring
them up to more modern standards. The code is the same as it was upon
removal, aside from making the journal not be colored.