Power Cell Bounty Fix (#21366)
Logistics bounties sometimes asked for a "Heavy Duty Power Cell" which
was an item that was not actually manufacturable by the ship, and could
only be obtained by stealing them from the APCs they spawned in. This PR
adds the ability for protolathes to manufacture them directly, just like
the other upgraded versions of APC power cells.