Dominia cape code tweaks (mainly: they all should now display in character setup) (#21363)
Changelog: 
> - rscadd: "The tribunalist medical cape can now have its mantle be
rolled up, like its non-medical tribunalist counterpart."
> - rscadd: "Alt-clicking on tribunalist capes now rolls their mantles
up."
> - bugfix: "All dominian capes should now properly display as attached
to clothing in character setup preview."

I have noticed that non-Zavodskoi dominian capes never seemed to display
as attached in the character setup, and attempting to solve that made me
notice how all the capes are poorly pathed and have vestiges of code
solutions made for their old sprite variations. So this PR, which was
made in like 30 minutes, _attempts_ to fix that.

I also have noticed that the tribunalist medical cape had sprites that
allowed it to be rolled up, but the code did not support that. Now it
does. I also added an alt-click functionality to roll them up, because
why not.